add missing multicodec mappings, render did:web documents
