+++ crypto_utils/lib/crypto_utils/keys.ex
@@ -46,11 +46,22 @@ defmodule CryptoUtils.Keys do
   @id_p256 {1, 2, 840, 10045, 3, 1, 7}
   @id_secp256k1 {1, 3, 132, 0, 10}
 
-  # Multicodec mappings
-  # @p256_code 0x1200
+  # Extra mappings not in Multicodec v0.0.2
+  @p256_code 0x1200
   @p256_prefix <<0x80, 0x24>>
-  # @secp256k1_code 0xE7
+  @secp256k1_code 0xE7
   @secp256k1_prefix <<0xE7, 0x01>>
+  @p256_mappings [
+    %Multicodec.MulticodecMapping{code: @p256_code, codec: "p256", prefix: @p256_prefix},
+    %Multicodec.MulticodecMapping{
+      code: @secp256k1_code,
+      codec: "secp256k1",
+      prefix: @secp256k1_prefix
+    }
+  ]
+
+  # Add these to our homegrown routine
+  @multicodec_mappings @p256_mappings ++ Multicodec.mappings()
 
   @doc """
   Generates a new random public-private key pair. `fmt` determines the
@@ -75,6 +86,57 @@ defmodule CryptoUtils.Keys do
 
     {pub, priv} = :crypto.generate_key(type, curve)
     {make_public_key(pub, curve, fmt), make_private_key({pub, priv}, curve, priv_fmt)}
+  end
+
+  @doc """
+  Extracts the public key from a "Multikey" verification method.
+
+  See `CryptoUtils.Keys.make_public_key/3` for details on the formats for the
+  returned key.
+  """
+  def extract_multikey(verification_method, fmt \\ :crypto_algo_key)
+
+  def extract_multikey(
+        %{"type" => "Multikey", "publicKeyMultibase" => multibase_value},
+        fmt
+      )
+      when is_binary(multibase_value) do
+    with {:ok, {pub, curve}} <- decode_multikey(multibase_value) do
+      {:ok, CryptoUtils.Keys.make_public_key(pub, curve, fmt)}
+    end
+  end
+
+  def extract_multikey(_, _), do: {:error, "not a Multikey verification method"}
+
+  @doc """
+  Decodes the public key from a "Multikey" verification method's
+  multibase value.
+
+  Returns `{:ok, {raw_public_key_bytes, multicodec_mapping}}` on success.
+  """
+  def decode_multikey(multibase_value) do
+    with {:ok, {public_key, :base58_btc}} <- Multibase.codec_decode(multibase_value),
+         {:ok, {raw_public_key_bytes, mapping}} <- multicodec_mapping_decode(public_key),
+         curve <- curve_from_mapping(mapping) do
+      {:ok, {raw_public_key_bytes, curve}}
+    end
+  end
+
+  @doc """
+  Decodes the public key from a "Multikey" verification method's
+  multibase value.
+
+  Returns `{raw_public_key_bytes, codec, multicodec_mapping}` on success.
+  Raises `InvalidPublicKeyError` on failure.
+  """
+  def decode_multikey!(multibase_value) do
+    case decode_multikey(multibase_value) do
+      {:ok, {_key_bytes, _curve} = tuple} ->
+        tuple
+
+      {:error, reason} ->
+        raise CryptoUtils.InvalidPublicKeyError, multibase: multibase_value, reason: reason
+    end
   end
 
   @doc """
@@ -203,5 +265,25 @@ defmodule CryptoUtils.Keys do
 
   def make_private_key(_, curve, _) do
     raise CryptoUtils.UnsupportedNamedCurveError, curve
+  end
+
+  defp multicodec_mapping_decode(prefixed_bytes) do
+    @multicodec_mappings
+    |> Enum.reduce_while({:error, "not found"}, fn %{prefix: prefix} = mapping, acc ->
+      plen = byte_size(prefix)
+
+      case prefixed_bytes do
+        <<^prefix::binary-size(plen), key_bytes::binary>> -> {:halt, {:ok, {key_bytes, mapping}}}
+        _ -> {:cont, acc}
+      end
+    end)
+  end
+
+  defp curve_from_mapping(%Multicodec.MulticodecMapping{codec: "ed25519-pub"}) do
+    :ed25519
+  end
+
+  defp curve_from_mapping(%Multicodec.MulticodecMapping{codec: codec}) do
+    String.to_existing_atom(codec)
   end
 end
