+++ did_server/lib/did_server_web/controllers/web_json.ex
@@ -1,4 +1,8 @@
 defmodule DidServerWeb.WebJSON do
+  def show(%{document: doc}) do
+    %{data: doc}
+  end
+
   def info(%{version: version}) do
     %{data: %{version: version, info: "did:web server"}}
   end
