+++ did_server/test/did_server_web/controllers/page_controller_test.exs
@@ -1,8 +1,8 @@
 defmodule DidServerWeb.PageControllerTest do
   use DidServerWeb.ConnCase
 
-  test "GET /", %{conn: conn} do
-    conn = get(conn, ~p"/")
+  test "GET /home", %{conn: conn} do
+    conn = get(conn, ~p"/home")
     assert html_response(conn, 200) =~ "Peace of mind from prototype to production"
   end
 end
